rename co2_emission to co2_equivalent_emission
