+++ src/climatevision/generator/agri2018/energy_demand/co2e_from_other.py
@@ -4,7 +4,7 @@ from dataclasses import dataclass
 
 from ...inputs import Inputs
 
-from ...common.co2_emission import CO2eEmission
+from ...common.co2_equivalent_emission import CO2eEmission
 
 
 @dataclass(kw_only=True)
