+++ src/climatevision/generator/agri2018/energy_demand/co2e_from_soil.py
@@ -2,7 +2,7 @@
 
 from dataclasses import dataclass
 
-from ...common.co2_emission import CO2eEmission
+from ...common.co2_equivalent_emission import CO2eEmission
 
 
 @dataclass(kw_only=True)
