+++ src/climatevision/generator/agri2018/energy_source/co2e_from_energy_use.py
@@ -2,7 +2,7 @@
 
 from dataclasses import dataclass
 
-from ...common.co2_emission import CO2eEmission
+from ...common.co2_equivalent_emission import CO2eEmission
 
 
 @dataclass(kw_only=True)
