+++ src/climatevision/generator/common/co2_equivalent_emission.py
